requirements.txt
Flask
Gunicorn    
numpy
pandas
Pillow==10.0.0
scikit-learn==1.3.0
keras==2.15.0  # Adjust the version as needed
tensorflow==2.15.0
matplotlib==3.7.2
# Add other dependencies as needed
